Correct width and alignment of containers
The containers of the main section and bottom section had different
widths, so their children weren't aligning properly with one another.
Don't know if using calc() was the best approach, but I think it looks
better now than it did before.

Also changed min-width of the highlighted article's image so that, at
the very least, it doesn't break the page on mobile (the header isn't
responsive yet so it will break anyways, though).